add snackbar notification when loading default providers
Issue:
The CC Catalog API adds (and sometimes removes) the image providers
available. Therefore, the extension loads the latest providers
using the /statistics/image endpoint of the API at the start of
 the popup session (same for the "Filters" tab in the options page).
If there is no internet connection or the API is giving some error,
extension loads from a "backup" list of providers.