remove feature of storing search results in localStorage
In the new UI of extension, this feature could be a reason for
confusion as due to this feature the filters were loaded in two contexts:
- previous search filters
- user filter preferences

In the old UI, it was easy to find out what filters are being set and in what context as filters were present in the home section. Now, they have their own section and therefore the users might not appreciate the need to switch to another section everytime they want to find out and confirm what filters are set.