adopt new storage schema for storing default filter settings
Until now, all the individual filter checkboxes were stored as
independent key in storage. The problem with this structure was that
when saving the user preferences more 60 writes were happening in the
storage at the same time. Therefore, the limit of 120 write operations
per minute was reaching more frequently.

In the new "schema", these checboxes are wrapped logically in their respective filter object. This would mean that only 8 writes per save will be needed.