enable navigating multiple image detail section in a single session
- Implemented a stack class for this. opening a new image detail section
pushed it's instance in the stack and closing pops it from the stack.
The bookmark or search gallery is not shown until all the instances of
image-detail section are not popped from the stack.
- Handled the case when user searches by an image tag. The expectation
then is to close the image detail section no matter how many instances
are stacked and show the results in the search gallery.