don't use backup sources when API connection fails
The number of sources in the API are regularly changing and also
sometimes sources are also dropped. So, having a hardcoded set of
sources, where some of them might not even be supported, does not makes
much sense.
If their is some error connection to the API while fetching the latest
sources, the user is notified.
Also, moved the checkInternetConnection() from searchModule.utils to
utils